+++ rust/guest/c/libc/include/stddef.h
@@ -1,0 +1,8 @@
+#ifndef KRAFT_STDDEF_H
+#define KRAFT_STDDEF_H
+
+typedef unsigned int size_t;
+
+#define NULL ((void *)0)
+
+#endif